Worked on Inventory.
I updated the UI so that it properly displays and assigns blank tiles into the inventory upon placement. I made a placeholder prefab for the thick rhombus tile so that we can see what the UI looks like. Then I went through the code to clean it up a bit, removing redundancy in a few different places and privatizing a bunch of attributes that did not need to be public. Yay! I don't think I broke anything this time!