Cleaned up the Merge and 'Improved' Tile Placement.
I moved all of the tile placement code into a script attached to the player called 'PlayerTilePlacement.cs'. I also worked on privatizing some stuff that didn't need to be public and did some small amount of commenting and redundancy fixing. Then I worked on the actual function that handles the tile placement and made it take a PenroseTile rather than just a thinRhombus. Now the math doesn't work so it's still always setting the tile played to thin Rhombus but less work down the line. I also made it return a bool value so that we can determine whether the tile is placed or not. This makes the player tile placement update the inventory appropriately! Yippee! Might do more but the battery on my laptop is dying!