More Commenting.
Okay, so. I think at some point, we did something and it broke the tile placement a little. Probably too far back to find where it broke again. We can talk tomorrow (I guess today) cuz we need to get this done and I have some thoughts on how to get that done.