Further tutorial work
I've added the custom font and art into the game, and built logic which determines strings to display for each tutorial stage.

I am close to implementing the checks for each tutorial stage, however I am struggling with it as there is not scoring system as of yet. Method TBD.