Level Things
I made a level manager that contains information about the level the player is playing. This includes if an initial tile should be pre-placed, if the player should earn a point for each tile laid and how many tiles should be randomly placed on the board. Additionally, I implemented each of these. Everything is connected to the new Level Manager object.