Reset Levels, Win Notification
Level Manager is now reset when main menu or level select menu is opened. This script is attached to the main camera cuz why not? I made a canvas that I attached the UI for winning. See LevelManagerActing.cs and the Exit action in PlayerInputs.cs for how to work with it. Basically, I just disable player actions except the Exit action and enable or disable the canvas that all the UI elements are connected to.