Levels Can Be Created.
Each button in the level select menu has a bunch of attributes that you can set to change the values of the Level Manager. This means that when the scene is loaded, the settings on the button will determine the settings of the level. Haven't tested it yet so let me know if you do and you find bugs. I have a suspicion about a specific possible problem.